update signature decorator tests
